+++ src/main/java/athena/exceptions/command/IllegalTimeModificationException.java
@@ -1,4 +1,4 @@
-package athena.exceptions;
+package athena.exceptions.command;
 
 import athena.ui.AthenaUi;
